refactoring
- Channel and Message init? from document
