+++ ChitChat02/DataManager/Firestore/models.swift
@@ -7,6 +7,11 @@
 //
 
 import Foundation
+import Firebase
+
+enum ChannelError: Error {
+    case name, timestamp
+}
 
 // MARK: Channel
 struct Channel {
@@ -14,6 +19,20 @@ struct Channel {
     let name: String
     let lastMessage: String?
     let lastActivity: Date?
+    
+    init?(from document: QueryDocumentSnapshot) {
+        guard let name = document.data()[Channel.name] as? String else { return nil }
+        guard let timestamp = document.data()[Channel.lastActivity] as? Timestamp else { return nil }
+
+        let id: String = document.documentID
+        let lastActivity = timestamp.dateValue()
+        let lastMessage = document.data()[Channel.lastMessage] as? String
+        
+        self.indentifier = id
+        self.name = name
+        self.lastMessage = lastMessage
+        self.lastActivity = lastActivity
+    }
 }
 
 extension Channel: Codable {
@@ -30,6 +49,20 @@ struct Message {
     let created: Date
     let senderId: String
     let senderName: String
+    
+    init?(from document: QueryDocumentSnapshot) {
+        guard let content = document.data()[Message.content] as? String else { return nil }
+        guard let createdTimestamp = document.data()[Message.created] as? Timestamp else { return nil }
+        guard let senderId = document.data()[Message.senderId] as? String else { return nil }
+        guard let senderName = document.data()[Message.senderName] as? String  else { return nil }
+        
+        let created: Date = createdTimestamp.dateValue()
+        
+        self.content = content
+        self.created = created
+        self.senderId = senderId
+        self.senderName = senderName
+    }
 }
 
 extension Message {
